% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryATAC.R
\name{queryATAC}
\alias{queryATAC}
\title{A function to query scATAC-seq datasets available in this package}
\usage{
queryATAC(
  geo_accession = NULL,
  author = NULL,
  journal = NULL,
  year = NULL,
  pmid = NULL,
  sequence_tech = NULL,
  score_type = NULL,
  has_clusters = NULL,
  has_cell_type = NULL,
  organism = NULL,
  genome_build = NULL,
  category = NULL,
  tissue = NULL,
  disease = NULL,
  metadata_only = FALSE,
  sparse = TRUE
)
}
\arguments{
\item{geo_accession}{Search by geo accession number. Good for returning individual datasets}

\item{author}{Search by the author who published the dataset}

\item{journal}{Search by the journal the dataset was published in.}

\item{year}{Search by exact year or year ranges with '<', '>', or '-'. For example, you can return datasets newer than 2013 with '>2013'}

\item{pmid}{Search by Pubmed ID associated with the study. Good for returning individual datasets}

\item{sequence_tech}{Search by sequencing technology used to sample the cells.}

\item{score_type}{Search by type of score (TPM, FPKM, raw count)}

\item{has_clusters}{Return only those datasets that have clustering results available, or only those without (TRUE/FALSE)}

\item{organism}{Search by source organism used in the study, for example human or mouse.
#TODO update docs}

\item{genome_build}{Return datasets built only using specified genome build (Ex. hg19)}

\item{category}{#TODO}

\item{tissue}{#TODO}

\item{disease}{#TODO}

\item{metadata_only}{Return rows of metadata instead of actual datasets. Useful for exploring what data is available without actually downloading data. Defaults to FALSE}

\item{sparse}{Return expression as a sparse matrix.
Uses less memory but is less convenient to view, recommended only if encounter memory issues with dense data. Defaults to FALSE.}

\item{has_cell_types}{Return only those datasets that have cell-type annotations available, or only those without annotations (TRUE/FALSE)}
}
\value{
A list containing a table of metadata or
one or more SingleCellExperiment objects
}
\description{
This function allows you to search and subset included scATAC-seq datasets.
A named list of scATAC-seq_data objects matching the provided options will be returned,
In cases where multiple matrices are available for each dataset, each matrix will be a seperate object within the list
if queryATAC is called without any options it will retrieve all available datasets.
This should only be done on machines with a large amount of ram (>64gb) because some datasets are quite large.
In most cases it is recommended to instead filter databases with some criteria.
}
\examples{

## Retrieve the metadata table to see what data is available
res <- queryATAC(metadata_only = TRUE)

## Retrieve a filtered metadata table that only shows datasets with
## cell type annotations and clustering annotations
res <- queryATAC(has_clusters = TRUE, has_cell_type = TRUE, metadata_only = TRUE)

## Retrieve a single dataset identified from the table
res <- queryATAC(accession = "GSE129785") #TODO maybe update this
}
\keyword{tumour}
